requirements.txt
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
plotly==5.14.1
nltk==3.8.1
transformers==4.29.2
torch==2.0.1 